staging: comedi: s626: use comedi_fc helpers to validate timer args
Use the comedi_fc helper cfc_check_trigger_arg_is() to validate the
trigger arguments when the source is TRIG_TIMER.

Use cfc_check_trigger_arg_min() to vaidate the scan_begin_arg when the
scan_begin_src and convert_src are both TRIG_TIMER.

All the arguments are unsigned int, change the local variable to an
unsigned int and rename it for aesthetic reasons.

Change the type of the first parameter to s626_ns_to_timer() to remove
the need for the casts.

Signed-off-by: H Hartley Sweeten <hsweeten@visionengravers.com>
Reviewed-by: Ian Abbott <abbotti@mev.co.uk>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>